Update to newest Rust

+++ src/datatype.rs
@@ -41,13 +41,13 @@ impl_data_type!(u8);
 impl_data_type!(u16);
 impl_data_type!(u32);
 impl_data_type!(u64);
-impl_data_type!(uint);
+impl_data_type!(usize);
 
 impl_data_type!(i8);
 impl_data_type!(i16);
 impl_data_type!(i32);
 impl_data_type!(i64);
-impl_data_type!(int);
+impl_data_type!(isize);
 
 impl_data_type!(f32);
 impl_data_type!(f64);
@@ -56,13 +56,13 @@ impl_data_type_ref!(u8);
 impl_data_type_ref!(u16);
 impl_data_type_ref!(u32);
 impl_data_type_ref!(u64);
-impl_data_type_ref!(uint);
+impl_data_type_ref!(usize);
 
 impl_data_type_ref!(i8);
 impl_data_type_ref!(i16);
 impl_data_type_ref!(i32);
 impl_data_type_ref!(i64);
-impl_data_type_ref!(int);
+impl_data_type_ref!(isize);
 
 impl_data_type_ref!(f32);
 impl_data_type_ref!(f64);
